+++ crates/kitsune_p2p/kitsune_p2p/CHANGELOG.md
@@ -4,6 +4,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## \[Unreleased\]
 
+- ShardedGossip can now accept `kitsune-mem://` urls as well as `kitsune-proxy://` urls for connections. [\#1033](https://github.com/holochain/holochain/pull/1033)
+
 ## 0.0.9
 
 - Fix rpc\_multi bug that caused all request to wait 3 seconds. [\#1009](https://github.com/holochain/holochain/pull/1009/)