+++ crates/kitsune_p2p/kitsune_p2p/CHANGELOG.md
@@ -3,7 +3,10 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/). This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## \[Unreleased\]
-- Check local agents for basis when doing a RPCMulti call. [#1009](https://github.com/holochain/holochain/pull/1009).
+
+## 0.0.10
+
+- Check local agents for basis when doing a RPCMulti call. [\#1009](https://github.com/holochain/holochain/pull/1009).
 
 ## 0.0.9